Check if input is missing, add name and examples for commander help
